se agrega funcion de decimales para las horas trabajadas

+++ database/migrations/2020_12_17_140403_gestion_solicitudes.php
@@ -31,7 +31,7 @@ class GestionSolicitudes extends Migration
             $table->date('fechaCambiada')->nullable();
             $table->date('fechaTermino')->nullable();
             $table->bigInteger('diasEjecucion')->nullable();
-            $table->bigInteger('horasEjecucion')->nullable();
+            $table->float('horasEjecucion')->nullable();
             $table->timestamps();
         });
     }
